+++ CHANGELOG.md
@@ -7,8 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- `Movie#stream()` to start streaming the movie to a `MediaStream`.
-- `Movie#getStream()` to get the current `MediaStream`.
+- `Movie#stream()` to stream the movie to a `MediaStream`.
 - Event constants (e.g. `Movie.Event.PLAY`).
 - Image and video layers' `source` properties now accept urls ([#153](https://github.com/etro-js/etro/pull/153)).
 - Movies, layers and effects have a new `ready` getter, indicating if they are ready to play.